+++ WebsitePhucKhao/Repositories/INhanVienPhongDaoTaoRepository.cs
@@ -3,9 +3,9 @@
 namespace WebsitePhucKhao.Repositories {
     public interface INhanVienPhongDaoTaoRepository {
         Task<IEnumerable<NhanVienPhongDaoTao>> GetAllAsync();
-        Task<NhanVienPhongDaoTao?> GetByIdAsync(int maNhanVien);
+        Task<NhanVienPhongDaoTao?> GetByIdAsync(long maNhanVien);
         Task AddAsync(NhanVienPhongDaoTao nhanVien);
         Task UpdateAsync(NhanVienPhongDaoTao nhanVien);
-        Task DeleteAsync(int maNhanVien);
+        Task DeleteAsync(long maNhanVien);
     }
 }
